requirements.txt
# local package
-e .

# external requirements
click
Sphinx
coverage
# awscli
flake8
python-dotenv>=0.5.1
langchain
huggingface_hub
datasets
apache-beam
tiktoken


#For parser
docx
python-docx
unstructured